server_app/app_grouper/requirements.txt
gunicorn
Flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7

torch
Pillow
numpy
torchvision
openai-clip
scikit-learn
tqdm
opencv-python